Implement easy/expert mode in config generation.
- Only show conf warning if has crypto passphrase
- Automatically upload crypto secrets when generated if the upload
  flag is set.  It will not upload any time it was generated to prevent
  issues with multiple systems

+++ ctutil/ct_certs.c
@@ -145,7 +145,7 @@ ct_get_cert_bundle(const char *user, const char *pass, char **xml,
 		e_free(&chunk.memory);
 		chunk.size = 0;
 	}
-	CDBG("login -> rc = %ld -> %s\n",rc, ru);
+	CDBG("login -> rc = %ld -> %s\n", rc, ru);
 	if (strcmp(ru, URL ACCOUNT)) {
 		rv = CT_CERT_BUNDLE_LOGIN_FAILED;
 		goto done;
@@ -170,7 +170,7 @@ ct_get_cert_bundle(const char *user, const char *pass, char **xml,
 		rv = -23;
 		goto done;
 	}
-	CDBG("get certs: rc = %ld sz = %ld-> %s\n", rc, chunk.size, ru);
+	CDBG("get certs: rc = %ld sz = %ld-> %s\n", rc, (long) chunk.size, ru);
 	if (!(rc == 200 && ru == NULL)) {
 		rv = -24;
 		goto done;
